intrinsics for first case in fill_error_matrix() function

+++ include/image_quilting.h
@@ -3,7 +3,7 @@
 
 #ifndef IMAGE_QUILTING_HEADER
 #define IMAGE_QUILTING_HEADER
-typedef int32_t pixel_t;
+typedef float pixel_t;
 #define PIXEL_T_MAX INT32_MAX
 
 typedef struct
